Krish-Learning/LangChain2/pyproject.toml
[project]
name = "langchain2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.0.21",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langserve>=0.3.2",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.35.0",
]
